Fix charainit param

+++ TarnishedTool/Models/ParamFieldDef.cs
@@ -11,7 +11,7 @@ public class ParamFieldDef
     public int? BitWidth { get; set; }
     public string DisplayName { get; set; }
     public int DefaultValue { get; set; }
-    public float? Minimum { get; set; }
-    public float? Maximum { get; set; }
+    public double? Minimum { get; set; }
+    public double? Maximum { get; set; }
     public int Offset { get; set; }
 }
